requirements.txt
Babel==2.8.0
coloredlogs==14.0
Flask==1.1.1
Flask-Caching==1.9.0
flask_CORS==3.0.7
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask_RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
iso4217==1.6.20180829
inflect==4.1.0
marshmallow-sqlalchemy==0.23.0
marshmallow_enum==1.5.1
passlib==1.7.2
psycopg2-binary==2.8.3
pyJWT==1.7.1
pylint==2.4.4
pytest==5.4.3
redis==3.5.3
requests==2.23.0
SQLAlchemy-Utils==0.36.6
webargs==6.1.0
